feat: move `LocalFeedImageDataLoader` and `FeedImageDataStore` to production
